pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mosaicai"
version = "0.1.0"
description = "A unified interface for multiple AI language models"
readme = "README.md"
authors = [{ name = "syukan3", email = "syun3ac@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
keywords = ["ai", "language model", "chatgpt", "claude", "gemini", "perplexity"]
dependencies = [
    "anthropic==0.33.0",
    "google-generativeai==0.7.2",
    "openai==1.40.3",
    "Pillow==10.4.0",
    "python-dotenv==1.0.1",
    "cryptography==43.0.0",
]
requires-python = ">=3.7"

[project.optional-dependencies]
dev = ["pytest==8.3.2"]

[project.urls]
Homepage = "https://github.com/syukan3/MosaicAI"

[tool.setuptools]
packages = ["mosaicai"]
